+++ tests/e2e/game/index.js
@@ -18,4 +18,5 @@ describe("E2E - 🎲 Game tests", () => {
     require("./pied-piper-infected.test");
     require("./unkillable-ancient.test");
     require("./ancient-infected.test");
+    require("./game-won-by-white-werewolf.test");
 });